+++ src/main/java/com/budgetmate/api/budgetmate_api/domain/expense/entity/Expense.java
@@ -37,12 +37,12 @@ public class Expense {
 
     private LocalDate expenseAt;
 
-    private int amount;
+    private long amount;
 
     private String memo;
 
     private boolean excludedSum = false;
-    public void updateExpense(LocalDate expenseDate, String memo, int amount, Long categoryId, boolean excludedSum) {
+    public void updateExpense(LocalDate expenseDate, String memo, long amount, Long categoryId, boolean excludedSum) {
         this.expenseAt = expenseDate;
         this.memo = memo;
         this.amount = amount;
